feat: add observable array
