--- Yaga/Utils/ObservableList.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: bd9882de1d464ad0bbde420a1b7874c1
-timeCreated: 1644300879
